#530: Move ItemCategoryControllerTests to ItemCategoryEndpointTests
