+++ Api/Xipona.Api.Endpoints.Tests/v1/Controllers/ItemCategoryEndpointTests/DeleteItemCategoryTests.cs
@@ -1,9 +1,8 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
-using ProjectHermes.Xipona.Api.ApplicationServices.ItemCategories.Commands.ModifyItemCategory;
+using ProjectHermes.Xipona.Api.ApplicationServices.ItemCategories.Commands.DeleteItemCategory;
 using ProjectHermes.Xipona.Api.Contracts.Common;
-using ProjectHermes.Xipona.Api.Contracts.ItemCategories.Commands;
 using ProjectHermes.Xipona.Api.Domain.Common.Reasons;
 using ProjectHermes.Xipona.Api.Domain.TestKit.Common;
 using ProjectHermes.Xipona.Api.Endpoint.v1.Controllers;
@@ -12,12 +11,12 @@ using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
 using System.Net.Http;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemCategoryControllerTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemCategoryEndpointTests;
 
-public class ModifyItemCategoryTests : EndpointCommandTestsBase<ModifyItemCategoryContract,
-    ModifyItemCategoryCommand, bool, ModifyItemCategoryTests.ModifyItemCategoryFixture>
+public class DeleteItemCategoryTests : EndpointCommandTestsBase<Guid, DeleteItemCategoryCommand,
+    bool, DeleteItemCategoryTests.DeleteItemCategoryFixture>
 {
-    public ModifyItemCategoryTests() : base(new ModifyItemCategoryFixture())
+    public DeleteItemCategoryTests() : base(new DeleteItemCategoryFixture())
     {
     }
 
@@ -43,27 +42,28 @@ public class ModifyItemCategoryTests : EndpointCommandTestsBase<ModifyItemCatego
         notFound!.Value.Should().BeEquivalentTo(Fixture.ExpectedErrorContract);
     }
 
-    public sealed class ModifyItemCategoryFixture : EndpointCommandFixtureBase
+    public sealed class DeleteItemCategoryFixture : EndpointCommandFixtureBase
     {
-        private ModifyItemCategoryContract? _contract;
+        private Guid? _id;
 
-        public ModifyItemCategoryFixture()
+        public DeleteItemCategoryFixture()
         {
-            PossibleResultsList.Add(new NoContentStatusResult());
-            PossibleResultsList.Add(new NotFoundStatusResult());
-            PossibleResultsList.Add(new UnprocessableEntityStatusResult(new List<ErrorReasonCode>()
+            PossibleResultsList.Add(new UnprocessableEntityStatusResult(new List<ErrorReasonCode>
             {
-                ErrorReasonCode.ItemCategoryNotFound
+                ErrorReasonCode.ItemNotFound
             }));
+            PossibleResultsList.Add(new NotFoundStatusResult());
+            PossibleResultsList.Add(new NoContentStatusResult());
         }
 
-        public override string RoutePattern => "/v1/item-categories";
-        public override HttpMethod HttpMethod => HttpMethod.Put;
+        public override string RoutePattern => "/v1/item-categories/{id:guid}";
+        public override HttpMethod HttpMethod => HttpMethod.Delete;
 
         public override async Task<IResult> ExecuteTestMethod()
         {
-            TestPropertyNotSetException.ThrowIfNull(_contract);
-            return await MinimalItemCategoryController.ModifyItemCategory(_contract,
+            TestPropertyNotSetException.ThrowIfNull(_id);
+            return await MinimalItemCategoryController.DeleteItemCategory(
+                _id.Value,
                 CommandDispatcherMock.Object,
                 ErrorConverterMock.Object,
                 CommandConverterMock.Object,
@@ -72,18 +72,18 @@ public class ModifyItemCategoryTests : EndpointCommandTestsBase<ModifyItemCatego
 
         public override void SetupParameters()
         {
-            _contract = new DomainTestBuilder<ModifyItemCategoryContract>().Create();
-        }
-
-        public override ModifyItemCategoryContract GetCommandConverterInput()
-        {
-            TestPropertyNotSetException.ThrowIfNull(_contract);
-            return _contract;
+            _id = new DomainTestBuilder<Guid>().Create();
         }
 
         public override void RegisterEndpoints(WebApplication app)
         {
             app.RegisterItemCategoryEndpoints();
+        }
+
+        public override Guid GetCommandConverterInput()
+        {
+            TestPropertyNotSetException.ThrowIfNull(_id);
+            return _id.Value;
         }
     }
 }
