+++ Api/Xipona.Api.Endpoints.Tests/v1/Controllers/ItemCategoryEndpointTests/ModifyItemCategoryTests.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
-using ProjectHermes.Xipona.Api.ApplicationServices.ItemCategories.Commands.DeleteItemCategory;
+using ProjectHermes.Xipona.Api.ApplicationServices.ItemCategories.Commands.ModifyItemCategory;
 using ProjectHermes.Xipona.Api.Contracts.Common;
+using ProjectHermes.Xipona.Api.Contracts.ItemCategories.Commands;
 using ProjectHermes.Xipona.Api.Domain.Common.Reasons;
 using ProjectHermes.Xipona.Api.Domain.TestKit.Common;
 using ProjectHermes.Xipona.Api.Endpoint.v1.Controllers;
@@ -11,12 +12,12 @@ using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
 using System.Net.Http;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemCategoryControllerTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemCategoryEndpointTests;
 
-public class DeleteItemCategoryTests : EndpointCommandTestsBase<Guid, DeleteItemCategoryCommand,
-    bool, DeleteItemCategoryTests.DeleteItemCategoryFixture>
+public class ModifyItemCategoryTests : EndpointCommandTestsBase<ModifyItemCategoryContract,
+    ModifyItemCategoryCommand, bool, ModifyItemCategoryTests.ModifyItemCategoryFixture>
 {
-    public DeleteItemCategoryTests() : base(new DeleteItemCategoryFixture())
+    public ModifyItemCategoryTests() : base(new ModifyItemCategoryFixture())
     {
     }
 
@@ -42,28 +43,27 @@ public class DeleteItemCategoryTests : EndpointCommandTestsBase<Guid, DeleteItem
         notFound!.Value.Should().BeEquivalentTo(Fixture.ExpectedErrorContract);
     }
 
-    public sealed class DeleteItemCategoryFixture : EndpointCommandFixtureBase
+    public sealed class ModifyItemCategoryFixture : EndpointCommandFixtureBase
     {
-        private Guid? _id;
+        private ModifyItemCategoryContract? _contract;
 
-        public DeleteItemCategoryFixture()
+        public ModifyItemCategoryFixture()
         {
-            PossibleResultsList.Add(new UnprocessableEntityStatusResult(new List<ErrorReasonCode>
-            {
-                ErrorReasonCode.ItemNotFound
-            }));
-            PossibleResultsList.Add(new NotFoundStatusResult());
             PossibleResultsList.Add(new NoContentStatusResult());
+            PossibleResultsList.Add(new NotFoundStatusResult());
+            PossibleResultsList.Add(new UnprocessableEntityStatusResult(new List<ErrorReasonCode>()
+            {
+                ErrorReasonCode.ItemCategoryNotFound
+            }));
         }
 
-        public override string RoutePattern => "/v1/item-categories/{id:guid}";
-        public override HttpMethod HttpMethod => HttpMethod.Delete;
+        public override string RoutePattern => "/v1/item-categories";
+        public override HttpMethod HttpMethod => HttpMethod.Put;
 
         public override async Task<IResult> ExecuteTestMethod()
         {
-            TestPropertyNotSetException.ThrowIfNull(_id);
-            return await MinimalItemCategoryController.DeleteItemCategory(
-                _id.Value,
+            TestPropertyNotSetException.ThrowIfNull(_contract);
+            return await MinimalItemCategoryController.ModifyItemCategory(_contract,
                 CommandDispatcherMock.Object,
                 ErrorConverterMock.Object,
                 CommandConverterMock.Object,
@@ -72,18 +72,18 @@ public class DeleteItemCategoryTests : EndpointCommandTestsBase<Guid, DeleteItem
 
         public override void SetupParameters()
         {
-            _id = new DomainTestBuilder<Guid>().Create();
+            _contract = new DomainTestBuilder<ModifyItemCategoryContract>().Create();
+        }
+
+        public override ModifyItemCategoryContract GetCommandConverterInput()
+        {
+            TestPropertyNotSetException.ThrowIfNull(_contract);
+            return _contract;
         }
 
         public override void RegisterEndpoints(WebApplication app)
         {
             app.RegisterItemCategoryEndpoints();
-        }
-
-        public override Guid GetCommandConverterInput()
-        {
-            TestPropertyNotSetException.ThrowIfNull(_id);
-            return _id.Value;
         }
     }
 }
